backend/requirements.txt
Flask
web3
python-dotenv
gunicorn
Werkzeug
setuptools
